approach is to get pool from helper
My concern here is that there might be some uncaught scenarios in which our main fails. Would it be better to not have our main return a result?

If this is a good approach i should probably merge the additional crate i made into the uniswap crate so that things are consolidated appropriately and thats where the pool struct is.