Bug fix process
First of all polars upgrade to 0.30.0 in main breaks things (even on main)
So the first fix here is to use 0.29.0

Then after that i noticed error discrepencies depending on the number of runs. So i started debugging for small number of runs first and fixed the shape error. Then i went up in number of runs untill the error came back and when we made it to number of runs > workers it breaks. Estelle helped debug.